MobileApp_Framework/requirements.txt
Appium-Python-Client~=2.7.0
selenium~=4.4.0
pytest~=6.2.5
allure-pytest
pytest-parallel
pytest-rerunfailures
pytest-sugar
pytest-json
pymysql~=0.10.1
sshtunnel~=0.1.5
requests~=2.24.0
invoke~=1.4.1
python-dotenv
pytest-html-reporter~=0.2.3
